fix #9
Looking closely at the bootloader update progress.

- After transferring the image, the nrf52 save new image in bank1, which
it will perform crc check

- If success then COPY new firmware from bank1 to bank0, erasing bank 0
and writing to bank 0

The issue is after sending the firmware and getting confirm that crc is
matched, nrfutil finished. And Arduino IDE re-open Serial which cause
nrf52 reset while in the middle of erasing bank 0 & writing bank 0. The
solution is simple calculate the delay which is time to erase + time to
write bank0

Worst Erase Page Time is 90 ms
Worst Write One Word Time is 338 us