Add SEGGER_RTT_TerminalOutBuffer()
This function is identical to SEGGER_RTT_TerminalOut(),
except that it sends the caller-specified arbitrary bytes
to the terminal.

This makes it much easier to use the multiple terminals
as a replacement for STDERR, STDOUT, etc., because the
function handles all the edges cases, and just needs
the formatted string to be provided.