Add support for the 7.2.0 version of S113 soft device (details below)
The S113 soft device removes BLE central support for a 54KB savings in
flash requirements over the S140 soft device (version 7.2).

This helps support for low flash devices (nrf52832 and nrf52833).  Also,
nrf52833 requires something softdevice 7.0 or later for bluetooth to work
at all.  So this enables the nrf52833 part to work with adafruit arduino
nrf52.

A few sections of the Bluefruit lib needed ifdefs to cope with missing
central support in the soft device.  I used BLE_GAP_ROLE_CENTRAL as
a conditional to guard those areas.

I'll also send in a related PR for the bootloader to add support for 7.2
series soft devices (required for BLE on nrf52833)